Move more stuff to types.py and implement audio timesource
